separate volume component definitions to each files
