+++ VMagicMirror/Assets/Baku/VMagicMirror/Scripts/Environment/PostProcessing/VmmAlphaEdgeVolume.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 2083cc9153d985e478b1755785a1d5d2
