+++ VMagicMirror/Assets/Baku/VMagicMirror/Scripts/Environment/PostProcessing/VmmCropVolume.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 24bd0c70c9954ee46a4c90490d21b013
